requirements.txt
pandas
requests
nsepython
yfinance
newsapi-python
python-dotenv
yagmail
